+++ supabase/migrations/20260801170000_m003_local_admin_role_membership.sql
@@ -1,8 +1,0 @@
--- Allow the local PostgreSQL administrator to exercise the trusted role matrix.
--- Application/browser roles remain unable to assume either privileged role.
-
-begin;
-
-grant app_workflow, app_migration to postgres;
-
-commit;